@@ -1,4 +1,9 @@
 // Cynhyrchwyd y ffeil hon yn awtomatig. PEIDIWCH Â MODIWL
 // This file is automatically generated. DO NOT EDIT
+import { lib } from "../models";
 
-export function Greet(): Promise<string>;
+export function Exit(): Promise<void>;
+
+export function InitFrontend(): Promise<void>;
+
+export function Refresh(): Promise<Array<lib.Device>>;
