@@ -2,6 +2,14 @@
 // Cynhyrchwyd y ffeil hon yn awtomatig. PEIDIWCH Â MODIWL
 // This file is automatically generated. DO NOT EDIT
 
-export function Greet() {
-  return window["go"]["main"]["App"]["Greet"]();
+export function Exit() {
+  return window["go"]["main"]["App"]["Exit"]();
+}
+
+export function InitFrontend() {
+  return window["go"]["main"]["App"]["InitFrontend"]();
+}
+
+export function Refresh() {
+  return window["go"]["main"]["App"]["Refresh"]();
 }
